Added digraphs to buildQuestionBank
Added some checks to finally allow digraphs to be properly supported. This was a bit of a process. Since, now that individual questions can be selected, there are multiple preferences that need to be selected to enabled digraph questions. One of the steps was ensuring the kana in the digraph questions were all the same (for lack of a better term) case, because normally the second character is a small character. The associated preference would be for the normal character. I had to write a special bit of code for that, since the toUpperCase method apparently doesn't work on Japanese characters. Then there's checking if both preferences associated with the digraph are enabled, and for those there are two preferences each, either the one for the full set, or the one for the specific character. I tried making the code as clear as possible, but there's likely more work to be done. For now, it works.