Changed display on QuestionSelectionDetail
Changed how the questions in the QuestionSelectionDetail box are displayed. It doesn't change how kana or kanji questions are displayed, but vocabulary questions will now have the english translation displayed, since it's more user-friendly. I basically just got the code to display the same thing it displays in the contents field of the QuestionSelectionItem object. When I first tried this, I just set the questions array to store the translations, but I forgot that the preferences are different, and need to be stored with unique and immutable keys, which can be different than what is displayed. So I had to set the QuestionSelectionDetail class to store two String arrays, one for the keys, and another for the names. To make things easier, the QuestionSelectionItem class stores an array of the associated questions, before sending it to the QuestionSelectionDetail box where the specific information is extracted when needed. In order to get all this working I also had to change a few of the methods in the Question classes to be public. It's a relatively minor change, overall.