Updated 'Selected' methods
Updated the six 'Selected' methods so they now work with the new design, and check question-specific preferences. As part of this, to support the digraphsSelected and diacriticDigraphsSelected methods, and prevent repeating code, I split off the code in the buildQuestionBank method, that checks the preferences for digraph questions, into its own method. That way I don't have to have the exact same code copied into those two 'digraphsSelected' methods, which I would've done. The code is pretty complicated anyway, so having it in its own method makes sense regardless.