Added additional check to getSelectedDigraphs method
When doing some debugging recently, I noticed a bit of a problem with the getSelectedDigraphs code. If one of the two question set preferences are set to false, it'll actually go through the process of checking each question, even though the end result will always be the same: nothing. So I added an additional check to the first if statement, ensuring that the code only runs if neither of the two question set preferences are set to false. This should speed things up. I also had to add null checks to the second if statement to prevent NullPointerExceptions, since a combination of null and false, which would otherwise cause said exceptions, will no longer be caught by the first if statement.