Added static modifier to constants
When I set up the converter to deal with the small characters in the digraph questions (1a8ebedd2650e8c2173f465b5c77f2130f8b7a9d), I forgot to set the two arrays it operates through as static. As a result, I'm pretty sure it creates multiple copies of the same array in RAM. So I fixed that.