Wrote code to handle swipe events
As I already explained in a previous commit (ff5e5694af6f787b687871a55025ef4836a360c9), the recent migration to ViewPager2 caused a few issues with the interface. I've been working to fix it for the past month, and just couldn't get anything to work; not because of a problem with the code, but because I couldn't find a way to get the code to actually run. Nothing I tried would work. I tried various event listeners, I tried overriding certain methods. I even tried creating a special view that would register and handle the movements while being placed in an overlay over top all the other objects. Nothing worked. The code just wouldn't run. I was pulling my hair out trying to figure out what was going wrong. I even found an official Android tutorial on how to handle gestures, and their example code wouldn't work either. It wasn't until browsing the relevant documentation on the OnTouchEvent method that I realized the problem. The internal views were picking up the events, handling them, and not allowing any other code to run. By disabling all other user-intractable code, including scrolling, I was able to, at the very least, test the code I wrote. There was a slight hiccup involving the fakeDragBy method not working as I expected in multiple ways. For a start, its logic regarding positive and negative values worked in the opposite way I expected. Secondly, it seems I had to provide the value relative to the last time the fakeDragBy method was called; and not, as I assumed, relative to when the beginFakeDrag method was initially called. But other then that, everything worked. But it only runs when scrolling and the long press function on the ReferenceCell are both disabled. There's still a bit more work to do, but at least I got something started.