Tweaked onTouchEvent logic
There was a problem with how the new page-turning code would work while backtracking. Basically, if one started swiping at the end of one of the sub-tabs to get to the next main tab, and then began swiping back, in the opposite direction, without lifting their finger; It wouldn't actually go back to the initial main tab, but instead the swiping back would affect the sub-tabs. This is mainly because most of the code was written with a misunderstanding of how the fakeDragBy method worked, as I explained in my previous commit (dc0100c2903971067e31c4c3e02adc97528fce29). So, with new understanding, I edited it so that movement is calculated using the new lastX variable, and the startX variable is merely used to determine which pager should be swiped. This allows for more fluid and intuitive operation. I also edited it so the logical operations determining if we're on the first or last tab in the nested pager are only executed once, when the swipe is initiated. This should save quite a few cycles.