Added condition to fake drag initialization
Added conditional statements to control the start and end of the fake drag of the main view pager object. This is mainly because, unless we're at the start or end of the nested view pager, there is no need to do any fake dragging on the main one. I mean, this is already handled by a conditional statement when the fakeDragBy is called, but I'm just applying that condition to whether the beginFakeDrag and endFakeDrag methods are called. I'm hoping this saves a few cycles.