Handled ACTION_CANCEL event
Looking through the MotionEvent documentation, I discovered that there was a 'cancel' event I was neglecting. And I realized that the occasional crashes I was experiencing with the message, "Cannot change current item when ViewPager2 is fake dragging" was likely caused by this. So, by adding it to the code so it runs the same code as the 'up' event, that ensures the whole thing is cleared out once the touch event ends, regardless of how it ends.