Streamlined code for fetching nested pagers
I never really liked how the code used for fetching the nestedViewPager was written. Just seemed a bit cumbersome. Also a bit redundant because most of it is a slightly-altered copy-paste of similar code from the ReferencePage class. So I decided to write some to-purpose methods specifically for fetching the correct nestedViewPager from the ReferencePager in the most efficient way possible. That mainly involved just cacheing the viewPager in each ReferencePage, as well as a record of the all the pages generated. I'm honestly quite happy with the end result. Should be more efficient than it was before.