Relinquished handling of subpager swiping
A problem I came to realize with my swiping code is that I should be handling as little as possible within it. Mainly so I don't break anything. Let the internal Views handle what they need to handle. I should only need to handle the page turns for the main view pager. So I decided to relinquish handling of the nestedViewPagers, and just allow their own code to operate as needed. This meant I no longer needed to cache the nestedViewPager objects. This not only simplified things a bit, it also meant the internal code works as it's meant to, and that I only need to handle behaviour that happens at the transition between two pages, nothing more.