Moved X-value caching
Because we don't need to store startX or lastX if we're not doing any page flipping, I decided to move the storage of those variables, as well as all operations, into the if statement that checks if we're at the start or end of the sub pager. Also, placed the page flipping code into an identical statement. This is because none of this code needs to be run if we're not using it. Now, the only code that is outside one of these statements is the code that checks those very conditions.