Smoothed out transitional points
In order to smooth out the otherwise janky transition while swiping between the nested pagers and the main pager, I rewrote the movement code. It's now designed to set the fake drag to only go as far as is intended, such that only movements that are on one particular side of the starting position are recognised. This is how it was always intended to work, but didn't whenever movement crossed over the starting line. These changes mean that only the distance traveled on the relevant side is now recognised, and there are no longer any gaps or overshoots, exactly as intended.