Simplified movement code
The movement code I wrote earlier (198d474bd1a92023b972c096d2b82f42178d1f6c), while effective, was quite bulky, and had a bit of redundancy in it. I found ways of simplifying it from 16 into just four lines of code. The basic logic and end result is identical to the previous code. I'm guessing it's slightly different in efficiency and flow. But the difference is negligible, and I believe, for the sake of simplicity, it's better this way. And the fact that the conditional statements are more complex is preferable over there being a multitude of them.