Restricted touch commands to specific space
It wouldn't make logical sense for page turning to happen when one is swiping above the main content body. So I set things up so there's a specific area where one has to swipe for the page turning to work. The hard part was recording what that area is. Because of how Android activities are built, there doesn't appear to be any standard method that runs automatically at start up *after* everything's drawn on-screen and all View objects have their dimensions established. Which is information I need to establish the specified zone. So I had to program it so the space is set up during first touch, since user interaction should only be possible once everything is drawn on-screen. It's the only way I could fathom to get it to work. It's not ideal though.