Restructured conditional logic
For the sake of efficiency and a simplified flow, I did a slight restructure of the conditional statements in the page turning code. Since two of the conditional blocks only run when isFirstItem or isLastItem is true, having that statement appear twice is a bit redundant, so I had both those blocks placed under a single instance of that one statement. This really changes nothing in terms of overall logic.