Fixed FragmentManager issue
Currently having a problem with the new ViewPager2. Every time I rotate the device, it crashes, every time, everywhere I use it. But weirdly, there are two different crashes happening. This fix only deals with one of them. The message I was getting was, "FragmentManager is already executing transactions". I didn't know what this meant, but I suspected it had something to do with the fact that I'm using a ViewPager2 within another ViewPager2. So, after a bit of googling, I found the solution was to use the childFragmentManager not the supportFragmentManager. And when I learned that, something clicked in my head. I remembered I had this problem before, and it had the same solution (6c973f1221b107a1e578d50b6a3f7729d91c34b6). But I didn't know how to implement it with the new FragmentStateAdapters, since the constructors don't accept FragmentManagers, but FragmentActivities. I found the solution though, by looking through the FragmentStateAdapter source code, where I noticed that the constructor that accepted FragmentActivities internally fetched the supportFragmentManager; and alongside that there was also a constructor that accepted simple Fragments, and that's the one that used the childFragmentManager. So, I switched it over, and now that crash is no longer happening. Other crashes are happening instead, but one thing at a time.