Added fixes to getItemId logic
I noticed a few problems with the getItemId method in the ReferenceSubsectionPager class. For a start, the if-else chain was missing an 'else', which breaks structured programming protocols. It wasn't a big deal, since the multiple return statements break that anyway, but that's for specific reasons, and it's the principle of the matter. I also noticed a problem with the itemIds for the kanji reference. I was storing the Locale hashCode in the itemIds, bit shifted 4 spaces to the left, to make room for a nibble representing which kanji file we're working with, and hoping it wouldn't result in overflow issues. But then, I realized that the Locale hashCode was an int, and the itemId was being outputted as a long. A long is twice the size of an int. So, I could instead redesign it, so the first half of the itemId is the Locale hashCode, and the second half is the number of the kanji file. And the brilliant thing about this is, not only is the design better and more logical, I no longer have to worry about overflow issues anymore. As part of this, I also assigned the value representing the bit shift of the Locale hashCode to a special constant, because I'm pretty sure I'm going to reuse it.