Added containsItem methods
I was getting an error with ViewPager2 on screen rotation, giving the message 'Design assumption violated.' Much like last time (0066c5254eb2a90d3cb69696ba519621ffaad829), I didn't know what it meant. So I did a bit of googling, and found the problem was mainly caused by my failure to read the documentation. Turns out, when you override the getItemId method, you're supposed to override the containsItem method as well, and I didn't realize. So, I quickly wrote the necessary methods. Part of this required writing a new method in the QuestionManagement class that fetches a question set number from the prefId. This is because I use the prefId to determine the itemId, and I need to reverse the process if I'm to determine if a given id exists.