Revert "Streamlined code for fetching nested pagers"
I was experiencing NullPointerExceptions on the reference screen every time I tapped anything after rotating the screen. I quickly traced the problem to the pageCache in the ReferencePager class. Seems it was getting erased on screen rotation, and trying to add items to or fetch items from the now empty List was causing issues. Now, I could've probably added some tests and conditions to work around the problem, but I realized this code that I'm having problems with was written to streamline the fetching of nested pagers. Trying to fix it that way would be the opposite of streamlining. So I decided instead to completely revert the commit (76fea047a02912d848f619eb17d9ba8830d190bc), and simply move the code it replaced to its own method. Not only because it looks better, but because otherwise I'd have to use that entire block twice.