Blocked swiping at extreme ends of the pagers
Because there's no reason to enable swiping of the main pager at the very beginning or very end of the pager set, I decided to disable that with a few additional conditions to the isFirstItem and isLastItem checks. Basically, isFirstItem is now only true if the nested pager is set to the first item, but the main pager isn't. And similarly, isLastItem is only true if the nested pager is set to the last item, but the main pager isn't. This prevents the swiping code from running unnecessarily in the simplest way possible.