Added touchSlop check
It's standard for apps to only count touch input as movement once the movement reaches a certain threshold, referred to as 'slop'. This is because without it, the human finger will wobble a bit anyway, and you don't want the app to consider that wobble as movement when the user doesn't intend that. So, I went through the trouble of integrating that feature into my swiping code. Basically I coded it so once that threshold is reached, it sets a flag, and we register all movement from that point forward until we get an ACTION_UP or ACTION_CANCEL event. That way, the movement doesn't get too janky; because it will otherwise.