Altered touchSlop handling
To allow more fluid movement, I changed how the touchSlop is checked, and the swiping boolean is triggered. Previously, if one swiped through the pager slowly, once the touchSlop was tripped, the page would jerk to the side. This is not the behaviour I wanted. After examining how the standard ViewPager2 works, I thought to mimic its behaviour. So now, once the slop is triggered, the point just before it's triggered becomes the new start point. The result is that the page appears to get pulled slightly. I think this looks a lot better, and smoother.