Blocked swiping while scrolling
So, because the app stops registering scrolling once it registers swiping in either direction, we ran into a problem. If the user starts scrolling on the pager edges where the swiping code is paying attention, but drifts too far to the left or right (which is a likely possibility), the scrolling will stop because the app assumes they're trying to swipe instead, and that means scrolling is suddenly cancelled, despite the fact that it's unlikely the user intended that. So, in order to stop this, I thought it would be a good idea, since swiping cancels all scrolling, that scrolling should cancel all swiping. I basically just set it up to keep track of the startY position, and disable all swiping once the Y-coordinate drifts outside the slop, because the scrolling will take over at that point. It was quite simple, if I'm honest.