Updated Android Studio
Updated to the newest version of Android Studio. As part of this, I not only allowed a few configuration files to just do their thing, I also updated the Android Gradle library to the newest version, and updated the target SDK to 30. I'm also getting alerts to remove the jcenter repository from my build files, and I noticed that things are moving on, and I'm expected to upgrade my project to Java 11, since Java 8 won't be supported with Android for much longer. But when I tried doing those things (removing jcenter and upgrading to Java 11), I started getting build errors, so it'll be a bit of work. For now, this is enough.