Replaced jcenter... as much as I can
I removed the jcenter repository from as much of the project as I can. I added mavenCentral to cover libraries not already covered by google and jitpack. However, the library for RikkaApps' FontProvider doesn't appear to be stored anywhere else. I thought I found it on some repository called 'Spring Plugins' but that turned out to be a bust. This basically forces me to keep the jcenter repository for that library alone. But I can only hope it gets uploaded elsewhere. Because once it does, I'll be adding that in a flash.