Partially reestablishing old compromise
Doing some testing I learned that the desugaring process that allows me to use java.time is not as effective as I presumed. The desugared classes are not playing well with the standard java classes. Which shouldn't really surprise me, since it's functionally just a backport, much like ThreeTen was. The good news is that it doesn't really change how anything works on Oreo or later (API >= 26), however on earlier models, crashes can happen if I feed a LocalDate class to a Formatter, for instance. Because of this, I had to revert my earlier commit (8561115ada2dc8794af0d4a98e02d5b1adad2167) and reestablish the old compromise. Thankfully, only on older systems.