Adjusted java.time workaround
After reimplementing the old compromise where I had to convert a LocalDate object to a long just so the String formatter would recognise it on old devices (71d1405241bef184fee0d796af519a85fe62944e), I realized I could just convert it to a java.util.Date object instead. And while java.util.Date is a terribly-implemented class, I'm barely interacting with it, and merely passing it directly to the String formatter, so it's fine, and much better than passing it a long.