Cleaned up submission
I found two problems with this pull request. For a start, I'm pretty sure they don't speak Germany in Spain, so I doubt there's a specific Spanish dialect of German, so I changed the locale code from German in Spain, to German in Germany. I'm just guessing, but I feel it's a safe guess. Second, I updated the copyright to the correct year and name, since the person who submitted this was not Laura C., and it wasn't written in 2019. Yes, the current year is 2021, not 2020. But it was written last year, in 2020, and I feel it's more appropriate to set the copyright to when the file was actually written, and no later. Don't know for certain if that's what's supposed to be done, but it's what I'm doing.