Merged KunYomiQuestion and OnYomiQuestion
In an effort to simplify things, I merged the two kanji sound question classes into a single class called... KanjiSoundQuestion. To make things more efficient, so I'm not storing unnecessary information, I decided to extend it off of the main Question class, and not the KanjiQuestion class. To differentiate kun'yomi and on'yomi, it's also storing the question type as a separate property. This should simplify things quite a bit.