Force sort of detail entries
I was having a bit of a problem with the entries on the ReferenceDetail box being out of order. Or at least, out of the intended order. I bounced around, trying different implementations of the Map interface, until I realized the simplest solution. I added a zero-width-joiner, which is an invisible zero-width character, at the start of an entry I wanted to force to the end; and I added a null character at the start of an entry I wanted to force to the start. Hopefully there won't be any major repercussions regarding this, but it's the simplest way I could do it. Another solution I thought of was to place a number at the start of each entry, and remove the first character when it comes time to display, but I felt that was a bit cumbersome. This is much simpler.