Moved CI reviews and fluff
I always thought the section on the CI services didn't fit in the app's readme. It's not really relevant to anything, just a bit of fluff. Especially the part where I just give my opinions on the services themselves. It's not necessary. But I didn't want to get rid of them, so I moved all that information to a separate file, with a link to said file replacing it in the readme.