Set up highlighting system
I noticed it was impossible to tell, from the QuestionSelection screen, which questions are selected without having to open the detail box. This didn't seem very user-friendly to me. So I decided I needed to fix that, and the simplest and most aesthetically-pleasing way to do that, I felt, was to highlight any questions that were selected in the contents field of the QuestionSelectionItem. This ended up being way more complex than I thought it would be. I knew the easiest way to do this would be by using HTML. I'm already using the fromHtml method, so it should be a cinch. The font tag is usable in this method, but the color property wouldn't accept anything but colour names or hex values; and the only hex values it would accept were RGB, not ARGB, which would include transparency. So I had to strip the transparency from the colour value. I decided, for efficiency's sake, to set the colour value once, when the QuestionSelection screen loads. Because it's not going to change unless one leaves the screen and comes back. So I don't need to set it in every object. The code for actually setting up the highlighted text ended up being a bit of a pain. I had to set things so the field reloads whenever anything changes. I also had to set special conditions for WordQuestions, since they're the only ones where the databaseKey, used for logs and preferences, is different from both the questionText, and what is displayed in the QuestionSelectionItem and QuestionSelectionDetail boxes. I also waffled about how to actually go about inserting the HTML. Initially I was going to set the highlighting in the contents property, and run the find and replace method to both set and unset the highlighting. But I realized how cumbersome that would be, so I instead decided to leave the contents property alone, and use it as a base to set the highlighting fresh each time the text box is reloaded. Then there was the matter of actually setting the highlighted questions. My first thought was to do standard string concatenation. But then, realizing that would result in either several string literals, or several named constants, I thought it was high-time I decided to try string formatting. It's likely more efficient, but something I never really bothered with in the past. In the end, I think it works well. I could probably clean it up in various ways. But now it works, and I'm happy with that.