Stopped improper highlighting
I found an error where, when 'Train' was selected in the vocabulary questions, not only would the 'Train' question be highlighted in the QuestionSelectionItem, but so would the word 'Train' in 'Bullet Train', which was not intended behaviour by a long shot. This also happens with some kana questions. In the extended kana set, the character 'ヴ' is not only a single question, but also a part of several other questions. In order to fix this, I figured simply encasing each question in null characters, which are zero width and invisible, would be a good way of indicating what is a complete question. That way, by setting up the find and replace appropriately, only the complete question would be highlighted. This prevents the misfiring I mentioned earlier.