Added scrolling and button to detail box
In order to improve usability, I added scrolling and an exit button to the detail box. This is because, with especially large question sets, they fail to all fit in the display area of the box, so the ability to scroll through them is necessary. Also, if the box fills the screen, there isn't really an obvious way to exit out, because tapping outside the box isn't really feasible. An exit button makes that easier.