Redesigned QuestionSelectionDetail
I redesigned the QuestionSelectionDetail box to work through the setMultiChoiceItems, because I realized that method existed, and was probably better than trying to build the views myself, individually. I had to do a lot of restructuring to get this to work, but I think it's a lot simpler, and more straightforward than it was before.