Moved a couple things out of loops
For the sake of efficiency, I moved a bit of code outside of the loops. This is code where the results wouldn't change between loops, so just storing the value once is much more efficient than running the same method over and over again, knowing I'll get the same result each time.