Fixed QuestionSelectionItem not updating properly
I found the reversal of the 'neutral' state of the QuestionSelectionItem wasn't being properly set in the event that the underlying checkbox object wasn't flipped. Whenever the item is checked on or off, by the onDismiss method of the QuestionSelectionDetail box, it's supposed to also set the preference for that entire group. But if the underlying checkbox object is already set to that state, than the preference won't be set. This actually results in a couple cascading errors. It means that the QuestionSelectionDetail box, if reopened, won't detect that it needs to set the parent object to 'neutral' upon any change to the selected questions. I'm not sure if I explained this issue correctly. But nonetheless, it has been fixed merely by specifically setting the preference in the event the checkbox's 'checked' state is not going to be flipped.