Changed highlight colour in light themes
In light themes, colorAccent is too dark to function as an effective highlighting colour. So I changed it to something that has a higher contrast with the normal text colour. I chose colorPrimary, which is the same colour as the title bar, which I'm not a huge fan of. But I am somewhat limited to the colours that are currently used in the theme. I guess I could figure out how to expand the colours available to me, but it's not an urgent priority, and so far, this works better than what came before.