Fixed how parent item is accessed
I found there were crashes triggered after the screen rotates when the QuestionSelectionDetail box is open. I quickly determined it was caused by the parent QuestionSelectionItem not being remembered as the box reloads. Trying to find a solution to this, I found I couldn't use the arguments Bundle because there was no method of storing a View. I tried setting the QuestionSelectionItem's id to something specific temporarily, and just looking up the parent item through that. But I found that once I set the id, I couldn't reset it, and setting multiple View objects with the same id would just result in the first item being the only one accessed. Eventually, I found the best solution would be to use the setTag method, and set every QuestionSelectionItem with a tag that is just that item's prefId. At that point, I'd just record the parent object in the QuestionSelectionDetail box by looking it up by said tag. This will be preserved between reloads, because the onCreateDialog method runs every time.