Changed SUBPREFERENCE_DELIMITER
In order to prevent any form of ambiguity, I decided to change the SUBPREFERENCE_DELIMITER to something that's not already used in prefIds. I figured a tilde would function the best in this circumstance. As part of this, I also decided to prohibit the use of tildes in prefIds by adding a new type to the xsd file. And while I was there, I also prohibited the use of spaces, since I don't really want spaces in prefIds, just for consistency reasons.