Fixed prefRecord
Before merging the finer-question-selection branch into this one (through the develop branch) (91028cecdb108a8f361072a3c1d85181979aa1d0), I knew there would be problems. I made significant changes to the code in both, and some of those changes were to the same code blocks, so they overlapped. And obviously, that caused merge conflicts. And those conflicts need to be fixed. I could've fixed them properly in the merge commit, but I decided instead to merge as rapidly as I could, and subsequently fix each error in a separate commit. I started with the prefId issues, since that's simpler. Since I added to what needed to be stored in one branch, and completely redesigned how everything was stored in the other, that obviously resulted in very obvious conflicts. Fixing it was easy, I just had to add the new prefs to the new design.