Made sound questions compatible with finer question selection
After the recent merge (91028cecdb108a8f361072a3c1d85181979aa1d0), and solving the initial conflicts (e40c0cdb16e70332c40c433aff4a94209e99aa09, 738546cecebefc9136850774b51dbeafaf210fad), I began testing and found that when KanjiQuestions are selected by group, the sound questions could be added if they are selected in the kanji_question_type preference. However, when KanjiQuestions are selected individually, the kanji meaning questions would be added, and nothing else, completely ignoring the kanji_question_type preference. I quickly realized why. The code that determines that is contained entirely within the addQuestions method. I was about to just move it to the addQuestion method, so it's run every time, but found the easiest way to do it would mean breaking structured programming, by having a return statement to cancel out the adding of the base KanjiQuestion when meaning questions are disabled. I didn't like that. So after thinking about it for a few seconds, I decided to just create a new private method for adding questions, and the public addQuestion method would be the one checking the kanji_question_type preference, and would just call the private method. It was the simplest way to do it without breaking structured programming.