Stopped continual resetting
I hated the fact that the caches in the QuestionBank class would be reset every time a question was added. Yes, this made sense, but it could happen like 50 times in a row when it only needs to happen once. So, I added a check to see if it was already reset. In that case, it wouldn't reset again. This should help with efficiency.