Replaced try-catch
I always knew it wasn't a good idea to use a try-catch as part of normal operation, like I did in the Function class' toString method. I intended to replace that eventually, but never got around to it, until now. I replaced it by adding a simple check to see if I can even access the FRACTION_CHARS array with the numbers I have. The rest of the method operates as normal.