Redesigned simplify method
I redesigned the simplify method to utilize a method I designed that searches for the greatest common divisor using an algorithm I found out about in my university programming class (which was originally designed by Euclid, apparently). This is partly because I thought that looping through prime numbers was a bit too restrictive, but also because once I found out about this algorithm I knew I had to implement it. I also moved the code that sets the whole number component to be within the simplify method, since it's no longer using recursion. As well as that, I decided to call the method in the constructors, and not the toString method, since I already removed the code for the whole number component from there.