Improved float conversion
Changed the conversion from floating point to a completely different scheme. Instead of converting to a fraction with a denominator that has a lot of factors, and then simplifying, I decided to just run through all the fractions with denominators up to 256, and check which of them are within a range of 2^-16. This is a lot more versatile, and a lot simpler. I also added tests to check even more fractions, including those as small as 1/256, to check for precision.