Made Fraction class functionally immutable
By setting all the Fraction properties to be final, I make the class properly immutable, as in cannot be altered, even accidentally. But this does mean there were a few additional changes that had to be made. For instance, I had to move the simplify code from its own method, to be within the constructor, since the simplify code actually alters the objects properties, which isn't allowed anymore, and had to be done in advance. I also had to replace one of the constructors, the one that took in a float value, with a new static method, fromFloat. This is because I wanted to call the constructor with the simplify code, but couldn't unless it was at the beginning of this constructor, and there was a lot of code that needed to be run before that happened. The only way around this was by converting it to a static method. I also had to replace all calls to that constructor. Thankfully, there weren't many.