Added additional conditions to fromFloat
Added two specific cases to the fromFloat method. One in the case the fractional part is less than 2^-16, in which case we just count that as zero. The other case is where it is less than 2^-16 away from 1, in which case, we just count that as 1. This is mainly for the sake of consistency, regarding how we treat any inaccuracies in floating point numbers. It would otherwise likely be handled by the fallback code, but it's better to handle it up front, that way it doesn't have to go through the nested loops.