Created fromFloatTest
There was a block of code in the toStringTest that was mainly written to test the accuracy of the fromFloat method. I decided to remove that and replace it with an improved version in it's own test. Rather than check the Fraction's value through the toString method, which was difficult to analyse, it checks the value by directly accessing the Fraction's properties. Now, in order to do this, I did need to set the properties to public, but they're also final now, so it's not really a big deal.