Added some math to fromFloatTest
To make the test a bit more rigorous, I added a bit of floating point math to the fromFloatTest. This now works to test how far a floating point can drift unintentionally, and how much I need to compensate for that. I also added a message to the assertThat statements, so I know where exactly in the loop the test fails, and it's not some kind of mystery.