Completing updates
After updating all the libraries there were a few compatibility issues I had to deal with, the biggest being the requirement of Java 11. I had to scour the settings menu to find the option to force the use of Java 11 over Java 8. I also updated the build file's compileOptions accordingly. Then there was the issue of the SDK. I updated it to the latest version, along with the build tools. But I did run into problems with the minSdkVersion. This was something I've been putting off for quite a while, but I finally decided to update it to something much more recent. Looking at all the programming hooks I put in for SDK compatibily, I decided to up it to version 22 (Lollipop v2), which is more than sufficient, covering 97.3% of devices. Anyone running older versions of Android can just go away. They're 3.7% of the market, I can't be supporting them forever. I also had to add an extra bit to the Manifest file, but that was minor. It just allows other apps to launch this app from the MainQuiz activity. Before now the true value was just default, now it has to be specified. Which is fine.