Replaced AsyncTask in MultipleChoicePad
AsyncTask was deprecated a while ago, but I had no idea what I needed to do to replace it. Also, I knew it would take a lot of work, and it wasn't really needed, so I put it off in favour of more important things. But recently, I ran across a short article on multithreading in Java, and it seemed simple enough, so I thought it was worth a shot. I decided to test it out on one of the simpler AsyncTask implementations, and it worked flawlessly. I'm going to hold off on the rest. But for now, this is a good start.