Replaced AsyncTask in MainQuiz
Continuing to replace the deprecated AsyncTask. With this one, I had to implement a method of cancelling the thread. I just used a flag, since I think that's the way I'm supposed to do it. I could've used the interrupt method, but I think that just pauses it, which would leave it lingering in memory, and that is not good. And besides, a flag is good enough for these purposes.